fix: there is no non-interactive powershell
